Sumerize p-value and adapt layout
I have added some option to the r markdowns to avoid to delete figure in the code
I have summerized the p-value in a tab using the coxph method in a loop. This can be discussed. The usefull covariates are now determined we can do the model.
I have reduce the n umber of page, we still have marging to reduce pages